Create Exercise-27 codetantra Input and Output.c
Problem Description:

Binita is playing chess. The game will be played on a rectangular grid consisting of N rows and M columns. Initially, all the cells of the grid are uncolored.



Binita's initial score is zero. At each turn, he chooses some cell that is yet not colored and colors that cell. The score obtained in this step will be the number of neighboring colored cells of the cell that Binita colored in this step. 

Two cells are neighbors of each other if they share a side between them. The game will end when all the cells are colored. Finally, the total score obtained at the end of the game will sum up of score obtained in each turn.



Binita wants to know what maximum score he can get. Can you please help him in finding this out?



Constraints:

1 ≤ N, M ≤ 50


Input Format:

The Only line of input contains two space-separated integers N, and M denoting the dimensions of the grid.



Output Format:

Print the output a single line containing an integer corresponding to the maximal possible score Binita can obtain.