Create Exercise-29 codetantra Input and Output.c
Problem Description:

Laasya bought a new volleyball in the sports shop. It looks medium size. She somehow found the radius of the sphere. But she would like to know the volume of that ball. 



Can you help him in finding the Volume of the ball? 



Functional Description:

Volume = (4.0/3.0) × π × r^3 

π = 3.14



Constraint:

1.00 ≤ r ≤ 5.00



Input Format : 

The only line of input has a single value of type float representing the radius of the ball. 



Output Format: 

Print the volume of the ball in a single line.